+++ CHANGELOG.md
@@ -11,6 +11,7 @@ in 1.x and 0.x versions.
   * Added cutlist real area for sheet good part
   * Added cutlist auto orient display
   * Added cutlist part orientation check 
+  * Added cutlist multiple layers check 
   * Added "more" leftbar submenu 
   * Added new extension icon